Sizes going across an IPC should be uint32.
BUG=164946

Review URL: https://chromiumcodereview.appspot.com/11472038

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@171944 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 0ab5fab4939150bd0f30ada8a4bf6eb0f69d66c1

+++ content/common/gpu/gpu_command_buffer_stub.cc
@@ -642,7 +642,7 @@ void GpuCommandBufferStub::OnRescheduled() {
     ReportState();
 }
 
-void GpuCommandBufferStub::OnCreateTransferBuffer(int32 size,
+void GpuCommandBufferStub::OnCreateTransferBuffer(uint32 size,
                                                   int32 id_request,
                                                   IPC::Message* reply_message) {
   TRACE_EVENT0("gpu", "GpuCommandBufferStub::OnCreateTransferBuffer");
@@ -658,7 +658,7 @@ void GpuCommandBufferStub::OnCreateTransferBuffer(int32 size,
 
 void GpuCommandBufferStub::OnRegisterTransferBuffer(
     base::SharedMemoryHandle transfer_buffer,
-    size_t size,
+    uint32 size,
     int32 id_request,
     IPC::Message* reply_message) {
   TRACE_EVENT0("gpu", "GpuCommandBufferStub::OnRegisterTransferBuffer");
